requirements.txt
absl-py==1.4.0
accelerate==0.19.0
aiohttp==3.8.3
aiosignal==1.2.0
anyio==3.5.0
appdirs==1.4.4
appnope==0.1.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.0.5
async-timeout==4.0.2
attrs==22.1.0
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.11.1
binaryornot==0.4.4
bleach==4.1.0
Bottleneck==1.3.5
brotlipy==0.7.0
certifi==2023.5.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.0.4
click==8.1.3
comm==0.1.2
contourpy==1.0.5
cookiecutter==1.7.3
cryptography==38.0.4
cycler==0.11.0
dataclasses==0.6
datasets==2.12.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
docutils==0.20.1
entrypoints==0.4
evaluate==0.4.0
executing==0.8.3
fairscale==0.4.13
fastjsonschema==2.16.2
filelock==3.9.0
flit_core==3.6.0
fonttools==4.25.0
frozenlist==1.3.3
fsspec==2022.11.0
future==0.18.2
huggingface-hub==0.15.1
idna==3.4
importlib-metadata==6.6.0
ipykernel==6.19.2
ipython==8.10.0
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.1.2
jinja2-time==0.2.0
joblib==1.1.1
json5==0.9.6
jsonschema==4.17.3
jupyter_client==7.4.9
jupyter_core==5.2.0
jupyter-server==1.23.4
jupyterlab==3.5.3
jupyterlab-pygments==0.1.2
jupyterlab_server==2.16.5
kiwisolver==1.4.4
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mistune==0.8.4
mpmath==1.3.0
multidict==6.0.2
multiprocess==0.70.14
munkres==1.1.4
nbclassic==0.5.2
nbclient==0.5.13
nbconvert==6.5.4
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
notebook==6.5.2
notebook_shim==0.2.2
numexpr==2.8.4
numpy==1.23.5
packaging==22.0
pandas==1.5.2
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pip==22.3.1
platformdirs==2.5.2
pooch==1.4.0
poyo==0.5.0
prometheus-client==0.14.1
prompt-toolkit==3.0.36
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.0
pycocoevalcap==1.2
pycocotools==2.0.6
pycparser==2.21
Pygments==2.11.2
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
python-slugify==5.0.2
pytz==2022.7
PyYAML==6.0
pyzmq==23.2.0
readme-renderer==36.0
regex==2022.7.9
requests==2.28.1
responses==0.13.3
restview==3.0.1
rouge-score==0.1.2
sacremoses==0.0.53
safetensors==0.3.1
scikit-learn==1.2.2
scipy==1.10.1
Send2Trash==1.8.0
sentencepiece==0.1.99
setuptools==65.6.3
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
stack-data==0.2.0
sympy==1.12
terminado==0.17.1
text-unidecode==1.3
threadpoolctl==2.2.0
timm==0.9.2
tinycss2==1.2.1
tokenizers==0.13.3
tomli==2.0.1
torch==2.0.1
torchvision==0.15.2
tornado==6.2
tqdm==4.64.1
traitlets==5.7.1
transformers==4.30.1
typing_extensions==4.4.0
Unidecode==1.2.0
urllib3==1.26.14
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.58.0
wheel==0.38.4
xxhash==2.0.2
yarl==1.8.1
zipp==3.15.0
